Don't extend precision of input parameters when they are machine precision
The previous implementation had no effect anyway as we were setting
MachinePrecision instead of . It may be better to remove the
warning altogether and quietly extend the precision of the input until the
output has the desired machine precision. However, this must be done before
computing the eigenvalue as the precision of the eigenvalue has an important
effect on the precision of the renormalized angular momentum.