Ensure that AccuracyGoal targets the accuracy of the final result
This requires us to include prefactors when checking accuracy of sums.
This commit implements this in the evaluation of the radial functions
and disables accuracy tracking when computing amplitudes. The same
sums are used with different prefactors so it is difficult to track
overall accuracy, and in any case the evaluation of the amplitudes is
quite quick as there are no hypergeometric functions to evaluate.